+++ tests/commands/load/test_load_panel_cmd.py
@@ -3,6 +3,8 @@
 from scout.commands import cli
 from scout.server.extensions import store
 
+from scout.demo import panel_path
+
 
 def test_load_panel(mock_app, institute_obj):
     """Test the CLI command that loads a gene panel"""
@@ -19,6 +21,39 @@ def test_load_panel(mock_app, institute_obj):
 
     # Test CLI by passing the panel 'OMIM-AUTO':
     result = runner.invoke(
-        cli, ["load", "panel", "--api-key", "not_a_valid_key", "--omim"]
+        cli,
+        [
+            "load",
+            "panel",
+            "--maintainer",
+            "john@doe.com",
+            "--api-key",
+            "not_a_valid_key",
+            "--omim",
+        ],
     )
     assert "OMIM-AUTO already exists in database" in result.output
+
+
+def test_load_panel_maintainer_not_in_db(mock_app, institute_obj):
+    """Test the CLI command that loads a gene panel"""
+
+    runner = mock_app.test_cli_runner()
+    assert runner
+
+    assert sum(1 for i in store.panel_collection.find()) == 1
+
+    # Test CLI by passing the panel 'OMIM-AUTO' - maintainer not in db!
+    result = runner.invoke(
+        cli,
+        [
+            "load",
+            "panel",
+            "--panel-id",
+            "panel2",
+            "--maintainer",
+            "noone@no.no",
+            panel_path,
+        ],
+    )
+    assert "does not exist in user database" in result.output